requirements.txt
azure-ai-projects>=2.0.0
azure-identity>=1.16.0
openai>=1.0.0
pydantic>=2.6.0
python-dotenv>=1.0.1
jsonschema>=4.21.1
requests>=2.31.0

azure-storage-blob>=12.19.0
azure-servicebus>=7.12.1

# Observability (opcional)
azure-monitor-opentelemetry>=1.4.0
opentelemetry-sdk>=1.23.0
azure-core-tracing-opentelemetry>=1.0.0b11

# Local parsing de .msg (opcional)
extract-msg>=0.50.0
